Convert to TypeScript. BREAKING: constructor and push/pull interfaces
Change-type: major
