--- .npmignore
@@ -1,2 +1,0 @@
-# Empty npmignore to avoid ignoring js files when publishing
-
