+++ package.json
@@ -2,27 +2,32 @@
   "name": "docker-progress",
   "version": "4.0.3",
   "description": "This module provides additional progress mechanics on top of dockerode",
-  "main": "index.js",
   "repository": {
     "type": "git",
     "url": "git@github.com:balena-io-modules/docker-progress.git"
   },
   "author": "Pagan Gazzard <page@balena.io>",
-  "types": "index.d.ts",
+  "main": "build/index.js",
+  "types": "build/index.d.ts",
+  "files": [
+    "build/"
+  ],
   "scripts": {
-    "lint": "balena-lint -e js --typescript --fix index.js",
+    "clean": "rimraf build",
+    "build": "npm run clean && npm run lint && tsc",
+    "lint": "balena-lint -e ts --typescript --fix lib",
     "test": "npm run lint",
-    "prepublish": "npm run test"
+    "prepublishOnly": "npm run build"
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@types/bluebird": "^3.5.30",
     "JSONStream": "^1.3.5",
-    "bluebird": "^3.7.2",
-    "docker-toolbelt": "^3.3.10",
     "lodash": "^4.17.15"
   },
   "devDependencies": {
-    "@balena/lint": "^5.4.1"
+    "@balena/lint": "^5.4.1",
+    "@types/dockerode": "^3.2.3",
+    "rimraf": "^3.0.2",
+    "typescript": "^4.2.4"
   }
 }
