+++ tsconfig.json
@@ -1,0 +1,17 @@
+{
+	"compilerOptions": {
+		"module": "commonjs",
+		"outDir": "build",
+		"noUnusedParameters": true,
+		"noUnusedLocals": true,
+		"removeComments": true,
+		"sourceMap": true,
+		"strict": true,
+		"target": "es2018",
+		"declaration": true,
+		"skipLibCheck": true
+	},
+	"include": [
+		"lib/**/*.ts"
+	]
+}
